moved proto > api directory to match standards. Also modified telegraf config, added consul systemd unit, and telegraf consul service.
