+++ etc/consul/consul.json
@@ -1,0 +1,6 @@
+{
+  "telemetry": {
+    "disable_hostname": true,
+    "prometheus_retention_time": "20s"
+  }
+}
